Use Single Buffer For Sending Packets
Prior to this a new heap allocation is made every time you
encode a packet to be sent on the socket.  This change allocates
a buffer the max size specified by the network settings and it
is used to dump the encoded bytes into before being written
by the socket.

I also cleaned up the receive buffer a little; while it wasn't
much, it seems to save a few extra asm instructions :)

I tested this with the server/client examples and it seemed
to work.